+++ postcss.config.js
@@ -1,5 +1,0 @@
-const options = {
-  debug: true
-}
-
-module.exports = require('./extends')([['cssnano', false]], options)